Updated for case: video has no subs after 1st run

+++ 2_save.py
@@ -35,24 +35,29 @@ def save(video_url, srt_url, page_url):
         directory = "{}/{}/{}".format(base_folder, dirs[0], dirs[2])
         path = "{}/{}.mp4".format(directory, dirs[3])
 
-        if os.path.isfile(path):
+        if os.path.isfile(path) and srt_url is None:
+            return
+        else:
             os.remove(path)
-
+        
         Path(directory).mkdir(parents=True, exist_ok=True)
 
-        if srt_path is None:
-            ffmpeg.input(video_url).output(path, codec="copy").run()
-        else:
-            (  # NOTICE: CPU usage will increase through thread queue size, so adjust to liking
-                ffmpeg
-                .input(video_url, thread_queue_size=2048)
-                .output(path, vcodec="copy", acodec="copy", scodec="mov_text",
-                        **{'metadata:s:s:0': "language=eng", 'disposition:s:s:0': "default"})
-                .global_args('-thread_queue_size', '512', '-i', srt_url) 
-                .run()
-            )
-            database.execute("UPDATE Videos SET file_exists = TRUE WHERE pageUrl = ?", [page_url])
-            database.commit()
+        try:
+            if srt_path is None:
+                ffmpeg.input(video_url).output(path, codec="copy").run()
+            else:
+                (
+                    ffmpeg
+                    .input(video_url, thread_queue_size=2048)
+                    .output(path, vcodec="copy", acodec="copy", scodec="mov_text",
+                            **{'metadata:s:s:0': "language=eng", 'disposition:s:s:0': "default"})
+                    .global_args('-thread_queue_size', '512', '-i', srt_url)
+                    .run()
+                )
+                database.execute("UPDATE Videos SET file_exists = TRUE WHERE pageUrl = ?", [page_url])
+                database.commit()
+        except:  # Connection lost
+            save(video_url, srt_url, page_url)
 
 
 for video in database.execute("SELECT * FROM Videos WHERE file_exists = FALSE"):
